Refactor: Rename gpt-pilot to samokoder across the project
