+++ core/templates/tree/node_express_mongoose/routes/authRoutes.js
@@ -46,7 +46,7 @@ router.post('/auth/login', async (req, res) => {
 router.get('/auth/logout', (req, res) => {
   req.session.destroy(err => {
     if (err) {
-      console.error('Error during session destruction:', err); // gpt_pilot_debugging_log
+      console.error('Error during session destruction:', err); // samokoder_debugging_log
       return res.status(500).send('Error logging out');
     }
     res.redirect('/auth/login');
